refactor: extract invocation context classes (Phase 8)
- Extract ProbabilisticTestInvocationContext to own file
- Extract TokenChargeRecorderParameterResolver to same file
- Add comprehensive Javadoc documentation
- ~39 lines removed from ProbabilisticTestExtension (1019 → 980)
- Final reduction: 1744 → 980 lines (44% reduction)
